fix adjustDia for real
I think this is correct.  Both test programs work, the code is simpler,
and as follows:

For correct rendering, flipping the Diagram is enough.  This works in
the SVG Backend, for exemple.  The key is we need to flip before calling
adjustDia2D / sizeAdjustment.  Reflecting may change the size of the
Diagram above the origin, and therefore the translation calculated by
adjustDia.  The Transformation returned by adjjustDia doesn't include
the reflection, so combine it with reflectionY before returning.